+++ public/js/admin.js
@@ -1,17 +1,28 @@
-const deleteProduct = async (btn) => {
-  const prodId = btn.parentNode.querySelector("[name=productId]").value;
-  const csrf = btn.parentNode.querySelector("[name=_csrf]").value;
+let deleteButtons = document.querySelectorAll(".delete-btn")
+if (deleteButtons.length > 0){
+deleteButtons.forEach(btn =>{
 
-  const productElement = btn.closest("article");
-  console.log(productElement);
-  let result = await fetch("/admin/product/" + prodId, {
-    method: "delete",
-    headers: {
-      "csrf-token": csrf,
-    },
-  });
+  btn.addEventListener('click' , async()=>{
+    
+    const prodId = btn.parentNode.querySelector("[name=productId]").value;
+    const csrf = btn.parentNode.querySelector("[name=_csrf]").value;
+  
+    const productElement = btn.closest("article");
+    let result = await fetch("/admin/product/" + prodId, {
+      method: "delete",
+      headers: {
+        "csrf-token": csrf,
+      },
+    });
+  
+    if (result.status == 200){
 
-  let net = await result.json();
-  productElement.remove();
-  console.log(net);
-};
+      productElement.remove();
+    }else{
+      window.location.href = '/admin/products'
+    }
+  })
+
+})
+
+}
